Split vite fixtures
