+++ packages/knip/fixtures/plugins/vite/vite.config.ts
@@ -2,9 +2,6 @@ import react from '@vitejs/plugin-react';
 import { defineConfig } from 'vite';
 
 export default defineConfig({
-  resolve: {
-    extensions: ['.js'],
-  },
   plugins: [
     react({
       babel: {
