+++ packages/knip/fixtures/plugins/vite3/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@fixtures/vite",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "@vitejs/plugin-react": "*",
+    "@emotion/babel-plugin": "*",
+    "vite": "*"
+  }
+}
